+++ changelog.md
@@ -4,6 +4,7 @@
 * LLT-3832: Upgrade python cryptography to 41.0.1
 * LLT-3855: Add "release.py" helper script, for easier release preparation
 * LLT-3813: Update trust-dns to 0.22.1 
+* LLT-3901: Don't log warn for optional analytics fix.
 
 <br>